backend/requirements.txt
django>=4.2,<5.0
psycopg2-binary>=2.9
django-environ>=0.10
djangorestframework>=3.14
django-cors-headers>=4.0
